Update c-sharp names to match operation Id's
csharp code samples names are updated to match relavant operation Id's of the calls so that code samples can be linked directly to documentation.
