+++ src/api-sdks/connection-api/clients/csharp/examples/CodeSamples/Samples/AddCrossSection.cs
