+++ src/api-sdks/connection-api/clients/csharp/examples/CodeSamples/Samples/GenerateWord.cs
