+++ src/api-sdks/connection-api/clients/csharp/examples/CodeSamples/Samples/SetBearingMember.cs
