+++ src/api-sdks/connection-api/clients/csharp/examples/CodeSamples/Samples/UpdateMember.cs
